PyShield/requirements.txt
pyyaml==6.0.2
fastapi==0.115.0
uvicorn==0.30.5
requests==2.32.3
redis==5.0.8
scapy==2.5.0
python-dotenv==1.0.1
pydantic==2.9.2
jinja2==3.1.4
python-multipart==0.0.9
